Moved shared steps into a different file, so they can easily be found
