+++ behave/features/steps/cla_select_digital_case_by_call_back.py
@@ -1,5 +1,5 @@
 from features.constants import CLA_CALLBACK_CASES, CLA_FRONTEND_URL
-from features.steps.cla_in_scope import assert_header_on_page
+from features.steps.common_steps import assert_header_on_page, wait_until_page_is_loaded
 
 
 @given(u'I am viewing a callback slot')
@@ -12,8 +12,6 @@ def step_impl(context):
     start_page_url = f"{CLA_FRONTEND_URL}/call_centre/callbacks/"
     context.helperfunc.open(start_page_url)
     assert_header_on_page("Cases", context)
-    import pdb
-    pdb.set_trace()
     callbacks = context.helperfunc.find_many_by_class("CallbackMatrix-slot")
     callbacks[0].click()
 
@@ -24,12 +22,24 @@ def step_impl(context):
     cla_callback_case = context.helperfunc.find_by_xpath(xpath_string)
     assert cla_callback_case is not None, f"Cannot find any cases created on CLA Public in the chosen callback slot"
     # find the reference of the case associated with this icon and use it for next step
+    context.callback_case_element_link = cla_callback_case
 
 
 @when(u'I select a case created on CLA Public from the callback slot')
 def step_impl(context):
-    raise NotImplementedError(u'STEP: When I select a case created on CLA Public from the callback slot')
+    # find and click on the case in the callback list
+    assert context.callback_case_element_link is not None, f"No callback case to link to"
+    context.callback_case_element_link.click()
 
+
+@step(u'I am taken to the call centre case details page')
+def step_on_case_details_page(context):
+    # check the url of the page
+    # will look like /call_centre/CASEID/diagnosis/
+    callback_case_ref = context.callback_case_element_link.text
+    page = f"/call_centre/{callback_case_ref}/diagnosis/"
+    wait_until_page_is_loaded(page, context)
+    assert_header_on_page(callback_case_ref, context)
